pipeline/requirements.txt
kfp==2.0.0b4
git+https://github.com/devoteamgcloud/pipelines#egg=google-cloud-pipeline-components&subdirectory=components/google-cloud
tensorflow==2.7.0
tfx==1.5.0
tensorflow-model-analysis==0.36.0
google-cloud-aiplatform==1.16.1
